pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "ncountr"
version = "0.1.0"
description = "Python pipeline for Nanostring nCounter data analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Zicheng Wang", email = "zw2595@cumc.columbia.edu"},
]
keywords = ["nanostring", "ncounter", "gene-expression", "bioinformatics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "scipy>=1.7",
    "statsmodels>=0.13",
    "matplotlib>=3.5",
    "pyyaml>=5.4",
    "click>=8.0",
]

[project.optional-dependencies]
crossplatform = [
    "scanpy>=1.9",
    "anndata>=0.8",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
all = [
    "ncountr[crossplatform,dev]",
]

[project.scripts]
ncountr = "ncountr.cli:cli"

[project.urls]
Homepage = "https://github.com/princello/ncountr"
Repository = "https://github.com/princello/ncountr"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
